updated syntax for to be more pedantic ... for now

+++ modules/common/common.h
@@ -10,7 +10,7 @@
 #define DIM5 5
 #define DIM6 6
 
-#define BASE_LENGTH 1.0                    /* vehicle base length in meters */
+#define BASE_LENGTH (float)1.0                    /* vehicle base length in meters */
 #define MIN_RUD_ANGLE (float)(-0.2 * M_PI) /* minimum rudder angle in radians */
 #define MAX_RUD_ANGLE (float)(0.2 * M_PI)  /* maximum rudder angle in radians */
